refactor: Change position data type from `float` to `double`
`double` increases precision. It should also technically be the default
data type, but hopefully my current float variables will be enough

Reference: https://softwareengineering.stackexchange.com/a/188725

+++ simulation/Calculator.cpp
@@ -4,58 +4,58 @@
 #include <iostream>
 #include <stdexcept>
 
-float Calculator::calcDistance(Star& s1, Star& s2) {
-    const std::vector<float>& pos1 = s1.position;
-    const std::vector<float>& pos2 = s2.position;
+double Calculator::calcDistance(Star& s1, Star& s2) {
+    const std::vector<double>& pos1 = s1.position;
+    const std::vector<double>& pos2 = s2.position;
 
     if (pos1.size() != pos2.size()) {
         std::cerr << "ERROR: pos1 != pos2 " << "(" << pos1.size() << " != " << pos2.size() << ")" << std::endl;
     }
 
-    float dist = 0.0f;
+    double dist = 0.0f;
 
     for (int val = 0; val < pos1.size(); ++val) {
-        const float displacement = pos2[val] - pos1[val];
+        const double displacement = pos2[val] - pos1[val];
         dist += displacement * displacement;
     }
 
     return std::sqrt(dist);
 }
 
-float Calculator::calcDistance(Star* s1, Star* s2) {
-    const std::vector<float>& pos1 = s1->position;
-    const std::vector<float>& pos2 = s2->position;
+double Calculator::calcDistance(Star* s1, Star* s2) {
+    const std::vector<double>& pos1 = s1->position;
+    const std::vector<double>& pos2 = s2->position;
 
     if (pos1.size() != pos2.size()) {
         std::cerr << "ERROR: pos1 != pos2 " << "(" << pos1.size() << " != " << pos2.size() << ")" << std::endl;
     }
 
-    float dist = 0.0f;
+    double dist = 0.0f;
 
     for (int val = 0; val < pos1.size(); ++val) {
-        const float displacement = pos2[val] - pos1[val];
+        const double displacement = pos2[val] - pos1[val];
         dist += displacement * displacement;
     }
 
     return std::sqrt(dist);
 }
 
-float Calculator::calcDistance(Node* s1, Node* s2) {
+double Calculator::calcDistance(Node* s1, Node* s2) {
     if (s1 == nullptr || s2 == nullptr) {
         throw std::invalid_argument("Calculator::calcDistance() : One or both nodes are nullptr.");
     }
 
-    const std::vector<float>& pos1 = s1->position;
-    const std::vector<float>& pos2 = s2->position;
+    const std::vector<double>& pos1 = s1->position;
+    const std::vector<double>& pos2 = s2->position;
 
     if (pos1.size() != pos2.size()) {
         throw std::invalid_argument("Calculator::calcDistance() : Dimension mismatch between position vectors.");
     }
 
-    float dist = 0.0f;
+    double dist = 0.0f;
 
     for (int val = 0; val < pos1.size(); ++val) {
-        const float displacement = pos2[val] - pos1[val];
+        const double displacement = pos2[val] - pos1[val];
         dist += displacement * displacement;
     }
